refactor: remove raw data related components
