--- src/components/ListRaw.tsx
@@ -1,102 +1,0 @@
-import { Box } from "@mui/material";
-import { ElectricityData } from "../models/electricity";
-import {
-  MaterialReactTable,
-  MRT_ColumnDef,
-  MRT_ColumnFiltersState,
-  MRT_PaginationState,
-  MRT_SortingState,
-  useMaterialReactTable,
-} from "material-react-table";
-import { useMemo, useState } from "react";
-import { useGetElectricityData } from "../apis/electricityStatistics";
-type Props = {};
-
-const ListRaw = ({}: Props) => {
-  //manage our own state for stuff we want to pass to the API
-  const [columnFilters, setColumnFilters] = useState<MRT_ColumnFiltersState>(
-    []
-  );
-  const [sorting, setSorting] = useState<MRT_SortingState>([]);
-  const [pagination, setPagination] = useState<MRT_PaginationState>({
-    pageIndex: 0,
-    pageSize: 10,
-  });
-
-  const {
-    data: { data, meta } = { data: [], meta: { totalRowCount: 0 } },
-    isError,
-    isRefetching,
-    isLoading,
-    error,
-  } = useGetElectricityData(columnFilters, pagination, sorting);
-
-  const columns = useMemo<MRT_ColumnDef<ElectricityData>[]>(
-    () => [
-      {
-        header: "id",
-        accessorKey: "id",
-        filterVariant: "range",
-      },
-      {
-        header: "Start Time",
-        accessorKey: "starttime",
-        filterVariant: "datetime-range",
-        Cell: ({ cell }) => {
-          return cell.getValue<Date | undefined>()?.toLocaleString();
-        },
-      },
-      {
-        header: "Production Amount",
-        accessorKey: "productionamount",
-        filterVariant: "range",
-      },
-      {
-        header: "Consumption Amount",
-        accessorKey: "consumptionamount",
-        filterVariant: "range",
-      },
-      {
-        header: "Hourly Price",
-        accessorKey: "hourlyprice",
-        filterVariant: "range",
-      },
-    ],
-    []
-  );
-
-  const table = useMaterialReactTable({
-    columns: columns,
-    data: data,
-    enableGlobalFilter: false, //turn off built-in global filtering
-    manualFiltering: true, //turn off built-in client-side filtering
-    manualPagination: true, //turn off built-in client-side pagination
-    manualSorting: true, //turn off built-in client-side sorting
-    initialState: {
-      showColumnFilters: true,
-    },
-    state: {
-      columnFilters,
-      pagination,
-      sorting,
-      isLoading,
-      showAlertBanner: isError,
-      showProgressBars: isRefetching,
-    },
-    rowCount: meta.totalRowCount,
-    onColumnFiltersChange: setColumnFilters,
-    onSortingChange: setSorting,
-    onPaginationChange: setPagination,
-    muiToolbarAlertBannerProps: isError
-      ? { color: "error", children: "Error loading data: " + error.message }
-      : undefined,
-  });
-
-  return (
-    <Box>
-      <MaterialReactTable table={table} />
-    </Box>
-  );
-};
-
-export default ListRaw;
