+++ src/pages/LandingPage.tsx
@@ -1,5 +1,4 @@
 import { Box, Typography } from "@mui/material";
-import ListRaw from "../components/ListRaw";
 import DailyStatisticsList from "../components/DailyStatisticsList";
 
 const LandingPage = () => {
@@ -7,8 +6,6 @@ const LandingPage = () => {
     <Box>
       <Typography variant="h1">Daily electricity Statistics</Typography>
       <DailyStatisticsList />
-      <Typography variant="h2">Raw Data</Typography>
-      <ListRaw />
     </Box>
   );
 };
