+++ src/types/electricityData.ts
